requirements.txt
Flask==2.2.3
flask_cors==3.0.10
keras==2.12.0
numpy==1.24.2
Pillow==9.4.0
tensorflow==2.3.0
